Year 2015 - day 19
This was the first one that stopped me in my tracks for a few days.

I foolishly attempted to solve it bottom-up with A* (using the editing
distance from the node to the target molecule as the heuristic); then I
gave the greedy top-down approach a go; and after that I attempted to
build an LR parser which turned out to be a rabbit hole (at least I got
a nice refresher).

But alas, I had been grazing the solution with my fingertips; using A*
top-down worked like a charm!