Year 2021 - day 17
I've struggled to solve this one.

I got the first one analytically just by realizing that once the vx
slows to a halt completely you can fire as high as you want, as long as
the probe is within the area on a given second. Then I discovered that
you can always hit the lower corner because after a certain speed, for
every unit increased in velocity there's precisely one unit of increase
in depth two ticks later; so it was just a matter of solving it for that
depth.

For the second part I did a bit of guesswork and a friend reminded me
that the two axis are independent (and I knew this in high school :/) it
was just a matter of calculating them separately and linking them with a
dictionary indexed by time.