Year 2021 - day 18
I couldn't quickly figure out how to do apply explode on a tree; so I
worked directly on the strings... It's pretty much unreadable and hard
to debug because of all the indexing with subtle offsetting. Luckily
there were a lot of samples in the problem description and I was able to
write a comprehensive test suite.