Year 2021 - day 20
Not a difficult problem, just some tricky details to consider...

At first I wanted to implemented it _smartly_ with an Inverted image
class that would perform the complementary set operations; but then I
would need to invert the algorithm as well and started to get
complicated to keep track of everything in my head.

In the end using boolean flags and XOR'ing them did the job.