Year 2021 - Day 21
Beautiful problem!

The main idea behind it is that many states are equivalent,
irrespectively of what was the sequence of events that lead to them.

Based on this we can store the frequency of each equivalent state
defined by (turn, position, score) and calculate the next states
frecuency by multiplying the previous step frequence by the relative
frequency of the event (dice rolls) that makes the transition.

A small optimization here is to pre-accumulate the aggregated
frequencies of all the dice rolls for the turn.

Each player states frecuencies are independent from the other player's,
so we can compute them independently and then calculate the aggregated
frecuency by multiplying their frequencies together.

The tricky bit is to consider that Player 1's winning state depends on
Player 2's state _for the previous turn_ because the game ends
immediately as soon as the max score is reached (no chance for the
second player to tie the game in the same turn).