Year 2022 - day 10
I had a bit of a struggle with this problem. My first hindrance was due
incorrectly assuming the processor had an execution pipeline (meaning
instructions could overlap).

On the second part I kept glitching the last column of the screen
because I needed to normalize the cursor to 0-39 in order for it to be
consistent with the modulus operator (initializing it to 0 instead of 1
as in part1 did the trick).