Year 2022 - day 12
A good old pathfinding problem! I copied and pasted the A* implementation
from a previous year.

The only subtletly is that I skipped through the part that said the end
position had an height of 'z'.

The second part could be optimized using Bellman-Ford, or doing the
search in reverse until reaching one 'a'; but the initial solution
proved fast enough to run it multiple times with different starting
points.