Year 2022 - day 15
I felt a bit overwhelmed when I was reading the problem description,
but it wasn't too bad in the end.

The first part was solved adding points to a set, but that wouldn't cut
it for the second part. The trick there was to search along the
perimeter of the area that each scanner covers, which is enough to
drop the complexity from quadratic to linear.