Year 2022 - day 16
First hard problem. I was very tired after my works xmas dinner, I
checked the problem description and when back to bed.

I attempted it in the afternoon and find a not-very-efficient solution
for the first part, with a search algorithm insipired in A*, it would
find the solution in ~3 minutes.

The second part works for the sample input instantly, but for the real
thing it's been running for hours with no luck.