Year 2022 - day 17
I had a bit of trouble debugging the first part, where I would assign
the shape to the slice of the board instead of adding it (so an empty
gap in the shape would overwrite an already settled rock).

I haven't solved the second part yet, the algorithm seems to work for
the example but does not find the correct solution for the real input.