Year 2022 - day 19
This was a tough cookie! When I read it in the morning I felt absolutely
discouraged, since I had been having trouble with the previous days...

But after the whole day minding my own business I got back home and
decided to give it a go. The first part wasn't that difficult, it was
just a matter of doing an exhaustive exploration of all the options with
some caching.

For the second part that didn't work. I was already in bed when I realized
that given that the factory can only produce one robot at a time, there
is no point in having more robots than what are actually needed to
produce the requirements to build one geode cracking robot per turn.

With that and a more efficient state representation (tuples instead of
dicts) the search could finish in a few minutes.

An extra optimization that I was considering was to group equivalent
states ignoring the amount of geodes and geode-cracking robots, but it
wasn't needed in the end.