Year 2022 - day 20
Finally a break! Today's problem was relatively easy, the first part
was very straightforward to solve, only being careful to see that when
inserting you don't need to advance your cursor if the element is
shifted forward.

For the second bit, I read over the part that specified the order in
which the elements should move should be the order they appear in the
original list... When I noticed that, it was easy to fix (and a bit slow
due the lookups, but still finished in a few seconds).