Year 2022 - day 25
Very easy problem for the last day, it is basically converting from one
base to another with a carry. I already solved a very similar one for
Google's Foobar where you had to balance two objects on a scale and you
only had a limited set of weights, so you would need to put them in the
empty scale side (add) or in the object's side (subtract).

Merry Christmas!