Year 2015 - Day 22
Another easy problem to solve with A*.

The trickiest part was to code the simulation behavior in a way that
wasn't too convoluted.

An object-oriented with classes approach might have helped to avoid
some code repetition (like checking effects are not already applied,
keeping track of spent mana, etc.), but given that the search works
by creating many states it seemed more appropriate to use a very
lightweight representation with tuples and functions.