Year 2015 - Day 23
A very easy one, I was expecting we should have to deduce loop
results to optimize, but wasn't necessary.

The only thing that delayed me was splitting the instructions in
tokens, I wrongly assumed \W was whitespace (instead of \s) and
that was removing the minus sign from negative jumps... Be careful
when using regular expressions!