Year 2016 - Day 1
My first problem with Julia.

A rocky start. On part 2 I thought it meant repeated turn position,
not any part of the path, so that took a bit of head scratching.

About Julia, I think it's going to annoy the hell out of me... Things
I did not like in my first contact:

- 1 based array indices (that was expected)
- Pushing arrays into Set's is not forbidden, and it behaves oddly if you
  mutate the array (checking for membership returns false, even with the
  very same Array object)
- Having to specify the size of integers seems backwards (but I
  understand it is because performance reasons)
- Slices cannot have implicit or negative boundaries, need to use "chop"
  with head and tail arguments, and it only works with strings
- No shorthand syntax for dictionaries (I can live without this)
- Need to specify global for assignments in a block in the top level scope,
  not just inside a function
- Why "in" can be used as an infix operator, but "mod" has to be called
  as a prefix function? Seems arbitrary...