Year 2016 - Day 2
Easy and fun. The second part required to think a bit about the
coordinates system, if you place the "7" at [0, 0], then clamping is
just any point with a Manhattan norm <= 2. Funny that the same concept
was used in day 1 too.

Mapping from the position to the key was a bit tricky. I wonder if
there's a simpler way...