Year 2016 - Day 7
New annoyance with Julia's API's. The order of arguments should be as
regular as possible to facilitate their usage, and for functions with
strings and regular expressions we have:

- replace(str, exp => subs)
- split(str, exp)
- eachmatch(exp, str)
- match(exp, str)

It would be nicer if the string argument came first in eachmatch and
match. Maybe there is a reason for this, but it escapes me at this time.