Year 2016 - Day 11
This one was challenging, and I learnt a bunch of interesting Julia
things along the way.

I tried to figure out a solution that is kind of similar to how the
towers of Hanoi are solved (basically decomposing the strategy into
smaller steps with less floors and less components), but couldn't think
of anything that worked.

I settled with A* search, and it only worked for the second part when
I did the normalization of equivalent components permutations; solved
sorting the pairs of chip/generators by floor.

Overriding the getindex and setindex! methods to use the array indexing
syntax was a nice touch, as well as using the Int values of the direction
enum to calculate the next floor.