Year 2023 - Day 1
What a start! The first problem and I was already banging my head
against the wall for a bit. Not a very difficult problem, but it was
surprising that there was a "gotcha" already for the second part.

Some day I should learn how to do overlapping regexes in Python, but
for today reversing the string and the expression to search the last
occurrence of the pattern was enough (even more efficient).