Year 2023 - Day 2
What is going on this year? XD This problem was kind of heavy on input
parsing. I took some extra time with the style to make people happy
when sharing it ^_^

I tried to wake up early, but didn't start right on clock; I was still
sleepy. I started reading the problem in bed around 6:02, then I got up
and turned on the laptop.