Year 2023 - Day 3
What is going on this year? XD The first few days are usuallly a warmup,
but this year they are a bit more challenging than usual.

There are two hard problems in computer science: cache invalidation,
naming things, and off-by-one errors. I've had some of off-by-one
errors today before getting today's problem right 0_o

I tried to simplify a bit the code for finding the numbers by leaving
the new line characters as a sentinel, but they would be counted as
symbols too... I ended up doing it the hard way and duplicating code at
the beginning of each line.