Year 2023 - Day 5
I was extremely tired this morning and took a while to understand part 1,
but it wasn't that hard at all. On the contrary, part 2 proved to be
much more challenging because all the mental effort I had to put into
visualizing ranges of numbers and performing boolean operations on them.

Had I have some piece of paper at hand, maybe it would have been easier,
but alas I am in a hotel room in Paris with my work laptop instead of the
comfort of my home... Dammit, I just realized I brought a notebook.

Anyway, using python's doctest was a great idea, I should do it more often.

The greatest time waster was realizing the range subtract needed to have
the ranges sorted to work.