Year 2023 - Day 8
This was a rewarding problem, not too easy, not too hard.

For the second part I initially went for the "brute force" approach, but
after the first week that's usually not an option anymore.

I was getting ready to do some permutations for calculating the minimum
cycle length if each start state would arrive to more than one end state
in its path, but luckily I realized that each start state only arrives
to a single end state. That simplified the solution a lot.

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
  8   00:09:18   1987      0   00:24:57   1169      0