Year 2023 - Day 10
Finally a hard one! The first part was already laborious, but the second
was fiendish. I started at 6:00 and finished around 12:45. I took some
breaks, but it was almost non stop.

My first approach was following along the pipe and do a mark & sweep
starting from the each adjacent on the inside, for that I would use a
normal vector. My first version used numpy, but it is more efficient to
just do lookups instead of vector multiplications.

The part that I got wrong was when an area is enclosed only by corners
winding on the counter direction. Very rare case, that I solved by
checking the two normals on corners rather than just the normal to the
movement from one pipe to the next.

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 10   00:46:58   3015      0   06:46:25   8725      0