Year 2023 - day 11
I was attempting to solve the first part doing the right thing, because
I had the hunch that for part 2 I would need it anyway; but I kept
getting incorrect results... So I did the brute force expansion.

When I had to face part 2, it took me a little bit to realize where
the mistake was: when iterating over rows and columns I was assuming
that they were ordered. So when in calculating the distance between a
a pair of galaxies, if the first galaxy had a greater value for row or
column than the second, it wouldn't do the loop for that dimension.

All I had to do to fix it was to iterate from min_row to max_row and
from min_col to max_col.

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 11   00:31:40   4083      0   00:49:14   4254      0