Year 2023 - day 12
The first part was small enough to be solved with brute force. The second
part however... It was very difficult.

I ended up with a recursive solution with memoization. It takes 15s to
run, so there's probably a better way to do it. Oh, fixed it, an extra
argument that I added for debugging mas messing with the memoization.

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 12   00:34:53   2313      0   08:50:48   7281      0