Year 2023 - day 14
Another "simulate and find the cycle" problem.

The tricky bit was, as always, finding the "off-by-one" error in the
calculation of the remaining cycles.

My simulation part was pretty inefficient, part 2 takes about 10 seconds
to run for the input, but we are optimizing total resolution time, not
runtime ^_^

I think a better way to do this would be to split the problem
representation in two sparse matrices, one for the fixed rocks and one
for the moving rocks. This way the process of comparing with the previous
states would be faster, since the part that changes is much smaller.

Also my way of finding the cycle is pretty inefficient, I'm doing an
equivalent of the "bubble short", scanning the whole thing and moving
rocks one position until they settle.

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 14   00:12:07   1591      0   00:42:36   1382      0