Year 2023 - day 14
I was unhappy with the performance of part 2, so I rewrote it to use
a sparse representation and do rotations instead of coding each tilt
as a separate function.

Doing the rotations also simplifies the code a bit because the sparse
representation is row-major; so iterating over columns would had been
a bit more cumbersome.

Execution time went from 10s to 0.4s