Year 2023 - day 15
Very easy day, a simple "follow the instructions" problem. The only
complexity is to understand what a very long problem statement really
wants with a sleepy head.

I glossed over the part that said that lenses were listed from front
to back, but luckily the sample wasn't evil and showed the error (if
it was in such a way that the focal lengths arrangement was symmetrical
it would have been a real head scratcher for me to find out).

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 15   00:05:28   1463      0   00:26:18   1698      0