Year 2023 - day 17
It was about time we got an A* search based problem! :D

I just discovered that the implementation of A* for Python that I had
borrowed long ago and kept reusing was horrible; its performance was
degrading the bigger the open set got, because it used a simple "min"
function to get the next candidate to explore. I fixed it using a heap,
and what a difference!

Part 1 using "min" took 50s, using a heap took 1s.
Part 2 using "min" took 9m29s, using a heap took 3.6s

Once again, the trickiest part (besides the A* optimization) was an
off-by-one error with the number of steps that you can take in a
straight line.

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 17   01:15:13   1826      0   01:35:11   1770      0

I need to go back to last year's elephants and valves problem, I think
it could have been taking forever to execute because of this uO_o