Year 2023 - day 18
I "brute-forced" the first part doing a flood-fill algorithm. That
approach obviously wouldn't work for part 2.

Luckily I remembered that one of my colleagues' friends solved day 10
(the pipes problem) using Gauss' formula (a.k.a. the shoelace formula).
The only tricky bit was to take into account that we are using full
blocks and not discrete points as the polygon boundary, so we need to
add the perimeter (half of it, because half of it is facing inwards).

I am still thinking about the odd + 1 to the area formula... I believe
it is where the initial and final square overlap when closing the loop,
but it is a bit early to think straight.

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 18   00:25:40   1234      0   00:38:10    539      0