Year 2023 - day 19
The input parsing was a bit on the heavy side, but still very regular...
I think I should start refreshing some context free grammar parsers,
for what might come next.

The first part was pretty straightforward. When I saw the description
for part 2, I breathed deeply, took my notebook and went to the kitchen
to prepare some tea.

As the tea brewed, so did my ideas to solve it, and in a few minutes I
had drafted a strategy to solve it in my notebook. It required to
explore exhaustively all possible rules, and keep track of which subset
of all the possible parts combinations got there.

As the possible values for each part weren't huge, instead of rolling
my own range algebra (I should implement this as a library at some
point), I used explicit sets and set intersection.

Still efficient enough to finish in 0.15s

I think this is the first time I got to use the newish `yield from`
expression. It makes the recursive generators way less verbose :D

Once more, the dreadful off-by-one error was lurking, and it took me a
while to debug the missing `+ 1` in the `>` condition ranges.

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 19   00:40:37   2700      0   01:49:23   2233      0