Year 2023 - day 20
Very challenging problem today! Simulating the first part was already
tough, because the BFS in this scenario wasn't trivial to implement
(you needed to keep track of the pulse information, not just the node).

The second part seemed super easy, until my code had been running for
over a minute and hadn't finished. At that point I lost hope about the
naive solution working, and went to have breakfast.

Then I plotted the modules graph, and realized that the target 'rx'
module was fed by a single Conjunction module, and that Conjunction
module didn't have any cycles itself... So I could calculate when its
inputs produced `high` values independently and then figure out when
these the high values for all of them would happen simultaneously.

Luckily, each of the inputs for the last Conjunction module cycled in
with a single `high` value at the end, so that simplified calculating
when would all of them be `high`: the minimum common multiple of the
lengths of the cycles; which happened to be just the product because
they had no common factors.

I have to admit that I tried a submission that was *too high* and then
noticed the off-by-one error in the cycles length XD

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 20   01:39:43   3026      0   03:18:11   2606      0