Year 2023 - day 21
Super hard problem.

For part 2, thankfully the input was nicely shaped, or it might have
been an even harder problem. I tried to solve it calculating the number
of full plots, plus the tips and the edges, but I couldn't quite get the
formula right.

I ended up calculating the first few values of the series where
`steps % 131 == 26501365 % 131` with the brute force algorithm from
part 1, and then extrapolate for the term with `n = 26501365 // 131`.

I hoped it would work because how the input was shaped, the cross and
rhombus clean areas left enough room for the gardener to "catch up" in
a straight line if detoured by a small rocks wall. The growth pattern
is just adding an outer layer of tiles on each cycle, so it should be
like a https://en.wikipedia.org/wiki/Centered_square_number with some
constant factors applied.

For my input it was 14613*n*n - 14479*n + 3592; but n_0 was not part of
the correct sequence, so I needed to add a +1 (off-by-one-error we meet
again! xDDD)

It worked, but I feel a bit dirty. I will have come back some other day
to solve it properly.

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 21   00:29:48   2903      0   10:50:57   4552      0