Year 2023 - day 25
And this is the last day for the year! It should have been an easy
problem, but if I had studied the minimum cut for graphs in college, I
had already forgotten it...

My first approach was to calculate all the minimum paths using
Floyd-Warshall, and the edges that appear the most should be the ones
to cut. This worked flawlessly for the test data, but there was no
chance an algorithm that runs in O(|v|³) would finish for a graph with
~1500 vertices.

I decided to plot the graph using GraphViz to see if there was any
pattern in the input that I could exploit... And then I saw the
beautiful SVG with the vertices clustered in two regions, joined by
three edges. I selected one region, cut it, saved the SVG and then
counted how many ellipses were in the new file. That was the size of
one island, I only had to calculate the other by subtracting it from the
total number of vertices and that was the solution! :D

As I want to have a more or less reproducible solution in this repo
(and day 24th was already hacky enough outsourcing to Maxima solving the
equations), I have implemented a solution that doesn't use any classical
min-cut or max-flow algorithm, but rather does a simulation based on
attractive (edges) and repulsive (vertex closeness) forces and then
does a k-means clustering of the vertices positions. I did it just one
dimension, because it is way more efficient and for our purposes the
results should be identical.

After tweaking a bit the parameters, it returns the correct result in
a very consistent way :)

Merry Christmas!