Year 2016 - day 15
I must confess that I always struggled with the chinese reminder
theorem. It was one of those things that wouldn't stick in my head...

But at least I could notice immediately that I needed it here!

I implemented it (cool thing about Julia, it has the `invmod` built in),
but it wasn't giving the correct answer. I did a second version using a
sieve, which yielded the same result. I did a brute force version, and
the solution was different (and correct).

I noticed that this solution was the cycle period minus the other
analytic solutions... Why was that? Oh, of course, because I was
calculating the congruence with the offsets *after* the open slot, but I
should have done it with the offsets *until* getting to the slot again!

For example, if I have a wheel of size 6 and offset (1+1=2), and a wheel
of size 5 and offset (1+2=3); solving the diophantine equation gives
offset 8, but I should have solved it for -2%6 (=4) and -3%5 (=2)!