Year 2016 - day 16
This is one of the problems where Julia has really shined...

The `BitVector` built-in type allowed to handle with convenience and
efficiency the data; the ability to broadcast an operator was really
handy to negate the strings and the matrix syntax to concatenate vectors
was very succinct (albeit having to use `false` instead of a 0 was a
bit annoying - pun intended).

I felt there were some hidden optimizations given the symmetry of the
way the bits were extended; but part 2 ran so fast I didn't think it was
worth pursuing it.