Year 2016 - day 19
This problem was easy to solve with linked lists.

I think part 1 at least could be solved with some smart use of modular
arithmetic... In each round you discard either odds or even positions
(depending on if the previous round started with an odd or even number
of elves), so you should be able to walk the process back from the bit
representation of the number of elves (which tells you if the round
was odd or even after halving the participants) :thinking:

For part 2 I don't see a very obvious way to optimize, since the
elimination pattern is a bit more complex.