Year 2016 - day 18
Something very strange has happened with day 18... I solved the problem,
and I remember creating a commit for it, but somehow the file was gone.

There wasn't even a git commit, where I wrote about how impressed I was
with Julia's performance, since I didn't need to optimize the part 2
(I even mentioned the same thing happening for day 16); but in case it
was needed it seemed clear that this was a case to check when a row
repeats and calculate the cycle value, etc.

Not even in the `git reflog` was there a reference to that commit.

Luckily vscode kept track of the file in the local history, so I didn't
have to write it again.