Year 2016 - day 21
The most annoying kind of problem, where the complexity comes from
following convoluted instructions.

Julia's one-based array indexing was particularly cumbersome for this
problem, but I didn't find too many off-by-one errors (just in the
reverse positions command).

Instead of deriving the formula for the inverse of the rotate based on
char, I just executed the direct transform for all characters and saw
where the character landed in the transformed.