Year 2016 - day 22
The first part was trivial, the second part required careful inspection
of the input to realize what was going on (although the first part
hinted it).

I solved it by hand first, with some hiccups due the usage of x,y
instead of i,j... Using x and y for matrices should be punishable.