Year 2016 - day 23
Remember that on day 12 I wrote in the commit message that I could have
optimized the execution by pre-parsing and doing loop analysis? Well,
this time it was necessary!

This was probably the most complicated AoC problem I have solved so far.
I guess there is a simpler way to solve it, but I did it the hard way.
It required to be careful to not make a mess, and still it was almost
400 lines of code.

For this case I feel like an Object Oriented approach would had been
a bit cleaner, since I could have grouped together the instructions and
reuse common behavior more easily. But still, the dynamic dispatch
worked well enough.