Year 2016 - day 24
Simple search problem. I noticed my BFS implementation was wrong, I was
using a stack instead of a queue... At least the results were correct,
because I was updating conditionally the distances if a shorter path
was found, but it was taking way too long because it would process
nodes multiple times.

The part of finding the minimum walks could probably be optimized, but
with only 7 nodes it wasn't really worth it.