Year 2024 - Day 07 (opt)
Given that all the operands are positive and the operators are
only additive, we know the sequence of values is always increasing.

That means we can do an early exit if the accumulated value is
greater than the calibration value.

We also know that concatenating always produces a value greater
than adding or multiplying, so we can also exit if the concatenation
of the accumulated value and remaining operands is lesser than the
calibration value.

Those optimizations reduced the execution time from ~13s to ~7s in
part 2.