Year 2024 - Day 16
Today's problem took some effort.

For the first part it was easier to track the score alongside the
state, so we could calculate the score increment easily, but that
caused the A* algorithm to keep repeated equivalent states.

For the second part we had to tweak our A* implementation to keep
track of the ancestors so we could reconstruct the paths that lead
to a solution, and also do an optimization in the path exploration
to avoid re-checking subpaths that were already covered.

A very didactic streaming episode :)