Year 2024 - Day 17
The first part was a simple "follow the instructions" problem, but
in the second we needed some ingenuity.

I started by trying to reverse-engineer the formula, but it was too
tricky for me. So when I started printing out stuff I discovered
that when the leading digits of the output matched the instructions,
the suffix digits (in octal) were increasing monotonically.

With that knowledge, we could "brute force" only the most significant
digits each time.