Year 2024 - Day 18
Very simple path-finding problem; without needing to make any
adjustements to a standard search algorithm.

For part 2, as we didn't need the distance, a simple flood-fill
algorithm to see if the start and end were connected was a bit
faster than A*.

The only tricky bit was the x,y vs i,j coordinate systems; I
swapped the indices at first and forgot to re-swap them for the
output.