Year 2024 - Day 24
The second part was quite tricky, and my solution seems very brittle.

I started by plotting the circuit. This revealed that it had a simple
ripple-carry adder structure, which made things easier.

Next, for each output bit, I simulated the circuit for all combinations
of the input bits at its position and the immediately previous. If
some of the outputs didn't match the truth table, then I knew there
was a problem with that part of the circuit.

Then I inspected the circuit graph and saw which gates where incorrect,
by identifying that the output bit should be the result of an XOR of
its level input bits and the carry from the previous level, etc.

When I suspected two wires were swapped, I manually updated the input
file and re-run the check, until the conflicting outputs were fixed.

After getting zero errors, I sorted the swapped wires and submitted
the solution.

Once I could relax, I improved the solution to check pairs of swap
candidates and re-run the check until one swap fixed the output bit.

The swap candidates were all the inputs of the current output plus
the inputs of the next level output... This is the brittle part, if
and input file swaps one wire with another wire that is a few levels
later my solution wouldn't work.